pyproject.toml
[project]
name = "monster-brawl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "IntensiveRock", email = "froggycham@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "click>=8.3.1",
    "ipykernel>=7.1.0",
    "ipython>=9.9.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pillow>=12.1.0",
    "rich>=14.2.0",
    "textual>=7.3.0",
    "textualize>=0.1",
]

[project.scripts]
monster-brawl = "monster_brawl:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
